+++ kitsune/runtime/opencilk/kitocilk.h
@@ -70,8 +70,8 @@ extern "C" {
 
 void __kitocilk_finalize(void);
 void __kitocilk_initialize(void);
-uint32_t __kitocilk_num_workers(void);
-int64_t __kitocilk_reduce_num_partials(int64_t n);
+uint64_t __kitocilk_num_workers(void);
+uint64_t __kitocilk_reduce_num_partials(uint64_t n);
 
 #ifdef __cplusplus
 } // extern "C"
